requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
numpy>=1.26.0
scikit-learn>=1.3.0
torch>=2.1.0
matplotlib>=3.7.0
scipy>=1.11.0
joblib>=1.3.0
vmdpy>=0.1
